Migrate default configuration to ~/.projectx/projectx.conf
The standard ProjectX configuration file is typically ~/X.ini, or an X.ini
file created in the directory wherever ProjectX.jar is executed.

This commit moves the ProjectX configuration file into a "dotfile"
directory (~/.projectx) which ensures it is hidden by default on Linux
systems. If this directory does not exist it is created.

If an X.ini configuration file is found in the user's home directory it
is automatically migrated to ~/.projectx/projectx.conf, as long as the
destination file does not exist.